requirements.txt
absl-py==0.8.0
albumentations==0.3.3
appdirs==1.4.3
attrs==19.2.0
black==19.3b0
Click==7.0
cycler==0.10.0
decorator==4.4.0
glog==0.3.1
imageio==2.5.0
imgaug==0.2.6
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
munch==2.3.2
networkx==2.3
numpy==1.17.2
opencv-python-headless==4.1.1.26
pandas==0.25.1
Pillow==6.2.0
pretrainedmodels==0.7.4
protobuf==3.9.2
pyparsing==2.4.2
python-dateutil==2.8.0
python-gflags==3.1.2
pytest-cov==2.8.1
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
tensorboardX==1.8
toml==0.10.0
torch==1.3.0
torchsummary==1.5.1
torchvision==0.4.0
tqdm==4.36.1
